LF-9420: drm/imx: imx93-ldb: filter out modes with unobtainable pixel clocks
The current implementation filters out modes based on pixel clocks
exceeding 80MHz. But, there are modes with pixel clocks below 80MHz
which produce nothing on screen because we cannot synthesize the exact
pixel clock from the parent clock, choosing a close enough value instead.
However, this makes the display reject the signal due to bad input
timings.

The fix for this would be to simply test, in the mode_valid() hook, if
we can synthesize the exact pixel and LVDS clock rates, unless there's a
panel connected.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>