arm64: dts: ls1046a: add SerDes nodes
Add the OF nodes for the 2 SerDes blocks on this SoC, and tentatively
also add the DCFG_DCSR region which will be necessary in the future for
RCW override (for dynamic SerDes protocol change). That is not currently
implemented.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
(cherry picked from commit 05de0ce609f55bc62e42e98b5b8aee9c8918e064 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>