net: fman: add support for c73 pcs
Commit a7c2a32e7f22 ("net: fman: memac: Use lynx pcs driver") has
created a silly precedent that we should list all possible PCSes for all
link modes in the device tree, as if it's a parade or something.

To be clear, the described problem (the 4 QSGMII PCSs are on the
internal MDIO bus of the first MEMAC) did not warrant the proposed
solution ("multiple PCSs in pcsphy"). Just a single pcsphy-handle, to a
PCS that's not on the internal MDIO bus of the given MEMAC, would have
been sufficient. The other pcsphy-handles, for interface modes other
than qsgmii, serve no purpose when the port is in QSGMII mode, because
it's implausible that there will be any dynamic switching between QSGMII
and anything else.

Anyway, here we are, needing to list 1000 PCSs in the device tree.
The Lynx PCS also supports backplane modes, and those aren't
characterized by a specific phy-connection-type (that is a phylib phy
concept, which may as well be absent with backplanes). So we use the
"c73" string to denote the phylink_pcs that will be used for clause 73
auto-negotiation.

Due to restrictions having to do with the legacy of U-Boot device tree
fixups, yes, we do need a specific phy-connection-type of either
"10gbase-kr" or "1000base-kx" to bypass some brain-dead U-Boot logic.
But we ignore it in the driver and pass PHY_INTERFACE_MODE_NA to phylink
anyway.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>