net: pcs: mtip_backplane: fix ineffective mtip_run_irqpoll_once()
cancel_delayed_work_sync() is the wrong API to use immediately after
schedule_delayed_work(), because it races with the scheduling of the
work item, and if it hasn't yet been scheduled, it won't be scheduled at
all, just canceled without ever running (resulting in a failure to
detect link loss).

What we want is flush_delayed_work(), which actually always waits for
mtip_irqpoll_work() to be called and finish its execution.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>