arm64: configs: move CONFIG_TSN out of defconfig and into lsdk.config
Lu Wei reports that after the tsntool kernel netlink code gained a
build-time dependency on CONFIG_NET_SCH_TAPRIO, it stopped being
compiled as built-in despite us having CONFIG_TSN=y.

The reason is because in the command "make defconfig lsdk.config",
the files "defconfig" and "lsdk.config" are evaluated sequentially.

First is "defconfig" coming from upstream, and this has
CONFIG_NET_SCH_TAPRIO=m, which forces its CONFIG_TSN dependency to not
be able to be built-in (built-in code cannot depend on modules).

We have "lsdk.config" which makes CONFIG_NET_SCH_TAPRIO=y, but this is
evaluated too late and it doesn't change CONFIG_TSN from its "m" value.

The actual problem that is seen is that the enetc driver itself also has
a dependency on CONFIG_TSN, so it also gets built as a module in that
circumstance. That is undesirable (inconvenient).

By moving CONFIG_TSN=y to lsdk.config, it gets evaluated after the
CONFIG_NET_SCH_TAPRIO=y line, and the tsntool kernel code can be
built-in.

Reported-by: Wei Lu <w.lu@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
(cherry picked from commit 79711209485d5083eb4b17c8251a1b6f99ec32a6 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>