LF-10374 arm64: dts: imx8ulp: add nvmem support for FEC
According to the fuse map of i.MX8ULP, the MAC address of FEC port is
stored in the fuse from the byte offset 0xAC to 0xB1 in big endian
order. So the FEC driver should have the capability to read its own
MAC address from the corresponding fuse area.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Alice Guo <alice.guo@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>