LF-10523-17 firmware: arm_scmi: clock: support clock denied flags
The firmware may export flags to indicate whether the clock
is allowed to set rate, set parent, enable/disable from the Agent.

If Agent is not allowed to enable/disable, directly return success.
If Agent is not allowed to set rate/parent, directly return -EACCES to
avoid SCMI RPC calls.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Acked-by: Nitin Garg <nitin.garg_3@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>