LF-10593-1 drm/bridge: synopsys: dw-mipi-dsi: Support hardware version 1.51
The hardware version of Synopsys DesignWare MIPI DSI controller embedded in
i.MX95 SoC is 1.51.  However, the version register value reads as 0x30313531,
which is shifted 8 bits at right hand comparing to the existing hardware
version value 0x31333100(1.31) and contains the value of 0x30 at the 8 MSBs.
So, specifically check the hardware version 1.51 to ensure the controller
is properly programmed.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>