LF-10611 arm64: dts: imx8mq-evk-rpmsg: fix CMA memory reservation fail
The current m4_reserved memory in imx8mq-evk-rpmsg.dts is actually
conflicted with the CMA reserved memory, so we will meet the following
errors when booting with rpmsg dtb:
OF: reserved mem: failed to allocate memory for node 'linux,cma': size 896 MiB

m4_reserved: m4@80000000 {
	no-map;
	reg = <0 0x80000000 0 0x1000000>;
};

Since the following two upstream patches which are already in 5.15
kernel, the ZONE_DMA size can be configured through DT's dma-ranges and
i.MX can DMA the whole 32-bit addressable range. so we can expand the
alloc-range to make the kernel be more easy to find the suitable
reserved memory.

commit 8424ecdde7df ("arm64: mm: Set ZONE_DMA size based on devicetree's dma-ranges")
commit d78050ee3544 ("arm64: Remove arm64_dma32_phys_limit and its uses")

Fixes: 4b31fe552b5e ("LF-1106 arm64: dts: imx8mq-evk: reduce cma reserved memory size to 896M")
Fixes: c889f4ecce1f ("LF-665-18 arm64: dts: imx8mq: make sure cma do not overlap dma zone")
Reviewed-by: Jason Liu <jason.hui.liu@nxp.com>
Tested-by: Guangliu Ding <guangliu.ding@nxp.com>
Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>