LF-10759: arm64: dts: power on MU2 upon loading the SOF firmware on QM and QXP
Since the SOF firmware uses MU2 to communicate with SCFW
for pin control and clock-related operations, the IP needs
to be powered on before the firmware starts its initialization.
Since the platform driver for i.MX8QM and QXP powers on all
of the resources specified through the "power-domains" property,
this commit also adds MU2A to the list so that the IP will be
powered on before firmware initialization.

Reviewed-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>