LF-10835-1 net: enetc: adjust base time if it is in the past
When we configure stream gate instance table or time gate scheduling
table, it is possible that the "base time" is a time in the past, so
we need to adjust the "base time", otherwise, add or update operation
on these two tables will fail. The formula for adjusting base time is
as follows:

	new_base_time = base_time + (N * cycle_time)

where N is the smallest integer so the resulting time is greater than
current time. And "cycle_time" is the sum of all the intervals of the
entries in the schedule.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
(cherry picked from commit bab3b9f2f6d1777e3c4a3bf366e8eba0fa7662e8 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>