LF-10853-03: media: nxp: add v4l2 subdev driver for csi pixel formatter
The CSI2 pixel formatter is a module which can uses packet info, pixel
and non-pixel data from the CSI2 host controller and reformat them to
match Pixel Link (PL) definition. In addition, optionally, it can either
route non-pixel data or one of pixel data type to another pixel link
virtual channel.

So add a V4L2 subdev driver for it since it's a standalone module and
need to export more controls to userspace.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>