LF-10844-1: arm64: dts: imx95: remove power-domains of NETC
Currently, the enetc driver does not support system suspend and
resume. And it causes the entire system's suspend/resume to fail
to work. Because the NETCMIX will be powered off when the system
enters suspend mode so that mdio bus can't be accessed and leads
to kernel panic.

For enetc driver, the system suspend/resume is a new feature, it
will take a long time to develop and test. To solve the issue of
kernel panic in a short time and make the enetc do not break down
the entire system suspend/resume, this patch is to remove all the
power-domains of NETC so that the NETCMIX will not be powered off
when the system enters suspend mode.

FWIW, the NETCMIX will be powered on by the uboot.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
(cherry picked from commit 1266d126e6a2737632932f63b7bf1b14c635338f nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>