LF-10938: clk: imx95-blk-ctrl: enable runtime pm for vpu_blk_ctrl
We want to enable the runtime pm for vpu, and the power domain
IMX95_PD_VPU is set for vpu_blk_ctrl and vpu and jpeg. vpu and jpeg has
enabled runtime pm, but the vpu_blk_ctrl doesn't, so we can't power off
the vpu power as required.

Enable the runtime pm for vpu_blk_ctrl, then vpu power is runtime enabled.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: TaoJiang <tao.jiang_2@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>