LF-10944-1 mmc: sdhci-esdhc-imx: set the RSTT for standard tuning
RSTT is the bit 28 of register SYS_CTRL(0x2C), USDHC add this bit
definition to reset the tuning circuit.

According to the RM, clear the execute_tuning bit from 1 to 0 in
AUTOCMD12_ERR_STATUS will also set this RSTT bit to reset tuning
circuit. But if the original execute_tuning bit is 0, it will not
trigger to set RSTT to reset tuning.

To make the code stronger, set the RSTT directlly to reset tuning
circuit.

Tested-and-reviewed-by: Sherry Sun <sherry.sun@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>