LF-10950: Revert "Revert "LF-10521 arm64: dts: imx93: Add sleep pinctrl for device node""
This reverts commit 25e213cf3e5061e5983c1a9a9c3b1d8612b589b5.

Previously, we reverted this patch to make sure no block for system
suspend/resume test due to eqos can NOT work in 1000M mode after resume.
actually, the pinctrl sleep support is not the root cause of this issue.
The root cause is that the EQOS driver itself does not handle the
pinctrl sleep state & phylink suspend correctly. the eqos driver
issue has been fix in:

commit a167e7b48fbf ("IMXBSP-101 net: stmmac: fix wrong place to call pinctrl state change")

so add the pinctrl sleep support back.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>