LF-11076 net: enetc: refine the implementation of PSFP
In the previous implementation, the operations of stream gate control
list table (SGCLT) were nested in the operations of stream gate instance
table (SGIT), which was very inflexible and difficult to expand. For
example, this implementation could not be applied to i.MX943 switch. So
we moved the SGCLT operations out of SGIT. In addition, this patch also
renames some structural variables of PSFP-related tables and adjusts the
order of creation and deletion of related tables, and fixes some potential
bugs.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
(cherry picked from commit 241815d643061ebbed6399f87b27c18a8a8c80cf nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>