LF-10957 neutron: Add new functions to reset firmware if neutron gets stuck
In some cases, neutron npu may get stuck, although this is rare.

This patch adds function to restart the firmware when it gets stuck.

we will continue to improve the firmware in case the neutron gets stuck and try to recover on the firmware side if it is hung.

As of now, this patch is required in case the firmware is failed to recover on its own.

Signed-off-by: Jiwei Fu <jiwei.fu@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>