LF-11281 drm/bridge: imx95-pixel-link: Do not support sharing next bridge between pixel links
It is likely that the same next bridge is found by two pixel links, e.g.,
pixel link0 finds MIPI DSI controller as the next bridge, while pixel
link1 finds it too by accident because a connected LDB channel is already
disabled dynamically by OF changeset.  So, explicitly check the next
bridges of the two pixel links and if they are the same, take it as an
unsupported case.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>