LF-11341 arm64: dts: imx95-Titan: Fix I2C address conflict and speed
The I2C address of typec controller is 0x50 which conflicts with
address of EEPROM on Verdin DSI to HDMI adapter. Once the adapter
is installed, typec can't work.

Toradex confirmed to remove R97 on carrier board to change typec
address to 0x52 to fix the issue. So update the typec address in
DTS. Also the Verdin board does not support 1Mhz I2C speed, have
to change the I2C7 speed to 400Khz in DTS

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Han Xu <han.xu@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>