LF-11363-2: arm64: imx8mp-evk-dual-balser: enable isi0 mem-to-mem feature
Enable m2m in dual basler cameras' dts file. Because of the hardware
design, just the isi0 can be used for m2m feature, but the isi0 capture
and m2m function can't be used at the same time. So if enable the isi0 m2m
function, the isi0 capture function should be disabled.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>