LF-11385 clk: imx95: correct blk ctrl for lvds csr
BIT5 is for LDB PLL clk input, not for output. If BIT5 got cleared, the
LDB PLL will not lock, so we drop it to not touch it as of now. Using
the default reset value 0.

And BIT0 is for phy div, so also add it.

Reviewed-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Nitin Garg <nitin.garg_3@nxp.com>
Tested-by: Ming Qian <ming.qian@nxp.com>
Tested-by: Sandor Yu <Sandor.yu@nxp.com>
Reported-by: Ming Qian <ming.qian@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
(cherry picked from commit 9ccbf0c8505864798b4916f8e28b6d057e974735 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>