staging: fsl_qbman: account for pre-initialized BARs in case of kexec
The QMan FQD and PFDR, as well as the BMan FBPR private memory regions
are configured in BAR registers that can be set only once per SoC reset.
In cases such as kexec, a second kernel might attempt to initialize the
QBMan block a second time without a full reset.

In this scenario, in the second kernel, we validate whether the
pre-configured private memory regions match the reserved memory regions
provided in the device tree.

We rely on the bootloader to reserve the memory regions and set the
addresses in the device tree, so all successive kernels use the same
regions.

If the system was previously configured, make sure the FQs are all reset
before use, buffer pools are drained, and skip initializing the PFDRs.

Additionally, zero out the PFDR/FQD memory regions before use, as
indicated by the manual, and disable interrupts during portal recovery.

Series inspired by the following changes for the upstream QBMan driver:
https://patchwork.kernel.org/project/linux-arm-kernel/list/?series=154265&state=%2A&archive=both
In particular, this mirrors upstream commit 97777078d695
("soc/fsl/qbman: Rework QBMan private memory setup").

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>