soc: fsl: guts: make it easier to determine on which SoC we are running
Add a new qoriq_die enum to be used as an index in the fsl_soc_die
array. A new fsl_soc_die_match_one() function is also added so that we
can directly determine if the SVR is a match with a specific DIE.
The SVR value read from the DCFG CCSR is also kept in the global soc
structure so that it can be accessed when needed.

These changes are useful because the guts driver will need to easily
determine on which SoC it's running in the next patches.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
(cherry picked from commit b87cc878c696671b6ba531005366d06b917e4e3f nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>