arm64: dts: ls1088ardb: describe the on-board TI retimer
Now that we have a driver for the on-board TI retimer, we can fully
describe it in DT and take advantage of it by referencing it from the
ethernet node.

With this patch in place, we are now able to dynamically reconfigure
(only at boot time, since no MOD_ABS GPIO connected to the cage) the
DPMAC.1 ethernet interface depending on the SFP module inserted.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
(cherry picked from commit a7f81738bfed97c07f2712be5d4e28d242bb485b nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>