LF-11558 drm/imx: dpu95-crtc: Disable ExtDst input src in shadow for top plane in ->atomic_begin()
Unlike ExtDst in i.MX8qm/qxp display controller, i.MX95 display
controller ExtDst can only connect with LayerBlends through it's
input interface or the input source is disabled.  For i.MX8qm/qxp,
display driver may connect ExtDst with ConstFrame when the plane
with the highest zpos(is top) is being disabled.  For i.MX95,
it turns out that ExtDst input source has to be disabled in this
case, otherwise, display controller will be in panic mode if the
fetchunit backing that disabled plane is migrated to the new
display controller and backs the plane with the highest zpos(is top)
on the new display controller.  So, disable ExtDst input source
in shadow in ->atomic_begin() by selecting the source to be
DPU95_LINK_ID_NONE(0, "disable" according to register specification)
and hope the shadow will take effect in ->atomic_flush().

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>