drivers: gpio: qixis-cpld: extend support also for the LS1046AQDS QIXIS CPLD
Extend the driver to also support the GPIO controller found on the
LS1046AQDS QIXIS CPLD.

For usecases in which the parent device (the QIXIS CPLD) does not define
a regmap for its memory area, create the driver's own regmap which will
get passed to gpio-regmap trough devm_gpio_regmap_register().
In this case, the offset inside of the regmap will be zero since the
regmap points directly to the desired register.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>