LF-11845-1 net: enetc: convert mailbox messages to new formats
For ENETC, it has the following use cases:
1. ENETC is completely controlled by the A core and completely driven
   by the Linux kernel or DPDK.
2. ENETC is completely controlled by the A core, but PF is driven by
   the Linux kernel, and VF is driven by DPDK.
3. ENETC is completely controlled by the M core.
4. ENETC PF is controlled by the A core, and VF is controlled by the
   M core.
5. ENETC VF is controlled by the A core (Linux kernel or DPDK), and
   PF is controlled by the M core.
So we need a unified mailbox message format, so that no matter it is
in the Linux kernel, DPDK, MSDK or other systems, we only need to
implement one mailbox mechanism. After discussions with DPDK, Auto,
BSP and MSDK team, Claudiu finally helped to unify the new format.
Many thanks to Claudiu for the help. So this patch is to covert the
existing messages in linux kernel to the new format.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
(forward ported from commit 8fed14f982f23d4dbec232ba0f6212026265f753 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>