staging: fsl_qbman: consolidate STATE_RETIRED processing with the others in qm_shutdown_fq()
The last bit of code for the STATE_TEN_SCHED, STATE_TRU_SCHED, STATE_ACTIVE
and STATE_PARKED states is duplicated for the STATE_RETIRED state. That is,
bringing the FQ out of service.

Use a "fallthrough" statement to avoid code duplication and possible
bugs caused by having different implementations (as the one that was
just fixed by the previous commit).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>