net: dpaa2-eth: MACSec data path support
This patch adds support for MACSec data path.

First of all, for a frame to target the MACSec controlled port of a
physical interface its hardware annotation needs to be setup with the
MCV and MCVV (qualifier for MCV) bit set to 1. What this means for the
dpaa2-eth driver is that for each Tx frame which needs to go through
MACSec (dpaa2_macsec_skb_is_offload) the associated FD had to have the
hardware annotation area valid and marked (dpaa2_eth_mark_macsec()) with
the two bits mentioned above.

On the Rx side, the software is notified only that a specific frame is
MACSec - DPAA2_FAS_MS - and not the specific Rx SC that it came from.
This means that the driver is forced to restrict the number of secure
channels to 1. Once the driver reads the DPAA2_FAS_MS bit, it uses a
metadata_dst structure to guide the skb towards the upper macsec
interface.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>