LF-12040-1 net: enetc: adjust the management of NETC clocks
The netc_prb_ierb driver is probed earlier than other drivers of
NETC, such as enetc, emdio and timer. So it's better to move the
system clock to netc_prb_ierb driver.
In addition, it's more reasonable to enable the reference clock
when opening the interface and disable the reference clock when
stopping the interface.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>