net: sdk_dpaa: use phylink_interface_max_speed() instead of local phy2speed[]
We can piggyback on phylink's phylink_interface_max_speed() which is
better maintained than our local array - for example we will need this
also for PHY_INTERFACE_MODE_XGMII.

We don't really care about making the MAC integration on the SDK driver
any better than it currently is. We just don't want to bring in the
phylink build-time dependency for now, so we move phylink_interface_max_speed()
as a static inline function in phylink's header, rather than an exported
symbol.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>