LF-12371 net: enetc: fix suspend/resume issue when eth port down
When an Ethernet port is down, the resume function will not re-initialize
the PCI-related registers of the Ethernet port. However, at this point,
the port has already been power-cycled, causing the port to be unusable
when up after the resumen.

To fix this, during suspend/resume, even if the port is not running,
these registers will be initialized as well.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
(cherry picked from commit 61fc886ef26aeaeb922feb1be76ba4f0a7b4a4c5 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>