LF-12372 ptp: ptp_netc: fix the issue of PTP synchronization
The NETC timer uses the NETC system clock as its clock source and
the rate is 333333333 Hz, so the clock period is about 3.000000003
ns. However, the NETC RM simplifies the clock frequency to 333MHz.
According to this value, the clock period is about 3.003003003 ns.
Therefore, the recommended fractional part of the period in NETC RM
is 0xc4ce07. And we also set this recommended value in the driver
based on NETC RM. But there is an issue when ENETC is used as the
master clock to synchronize the clock with FEC, it will cause the
clock offset to be a bit big, is about ±200ns and cannot converge
to ±100ns, because the accumulated value of the ENETC clock counter
is larger than the actual time. So we need to set the fractional
part of the period correctly.
In addition, the max_adj was set to 1000000 (in parts per billion)
before, so the max adjustable offset of the period is about ±0.003
ns. If the NETC timer still runs faster than the peer (master clock)
after adjusting the max offset of the period, PTP synchronization
will not be completed because the period cannot be shortened anymore.
Therefore, adjust the max_adj to increase the max adjustable offset
of the period.

Fixes: 9eb53327f286 ("LF-10640-6 ptp: add NETC PTP driver support for i.MX95")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>