MLK-26186: media: os08a20: Fix V4L2_CID_ANALOGUE_GAIN
The os08a20_set_gain() function was inherited from imx8mp,
and it is specific to the VSI ISP, which is expecting a total
gain of 1024 for 1x gain, with the analogue and digital gain
being combined.

For imx95, add a new function and let the control update only
the analog gain registers with the gain code in sensor format.
This covers linear mode (gain for long exposure).

Fixes: 4892da9340be ("MLK-26186-5: media: os08a20: Add analog gain and exposure controls")

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>