LF-12764 net: enetc: fix issues when deleting VLANs or MACs
Based on the VSI-PSI message protocol, the PF should return a NOT_FOUND
error if delete inexistent MACs or VLANs. However, the current driver
ignores this error, so add checks for MACs and VLANs to be deleted.

Another issue is that the entry index is not reinitialized after the
table entries are deleted, which causes the index to exceed the maximum
value when adding new table entries, so the entries cannot be added.

Fixes: 8fed14f982f2 ("LF-11845-1 net: enetc: convert mailbox messages to new formats")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
(cherry picked from commit 808508113971dab329de4621b77d8c8ade8e5df0 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>