LF-12615-2 net: enetc: add devlink to reallocate the number of rings
Add devlink parameter to allocate the number of rings for PF and VFs
of the ENETC port. The parameter is a hex string and supports max 15
VFs (it depends on hardware capability). Each VF occupies 2 bytes of
the string, so the max number of rings of per VF can be set to 255,
but it also depends on hardware capability.

Numerically higher VFs have lower bytes of the string, as an example,
if one ENETC supports 4 VFs, and the string is set to "1020304" or
"01020304", then the ring allocation for each VF is shown below.
VF0 = 0x04
VF1 = 0x03
VF2 = 0x02
VF3 = 0x01

Note that the number of rings of PF is not explicitly presented in the
string, because the remaining number of rings is reserved for PF.

Examples:
1. The ENETC has 2 VFs and 8 rings, then set the number of rings of per
   VF to 2, and the number of rings of PF is 4.
   $ devlink dev param set pci/0002:00:10.0 name si_num_rings cmode \
     driverinit value 202

2. Then run devlink reload command to apply the new configuration.
   $ devlink dev reload pci/0002:00:10.0

3. Get the value of si_num_rings
   $ devlink dev param show pci/0002:00:10.0 name si_num_rings
   pci/0002:00:10.0:
   name si_num_rings type driver-specific
     values:
       cmode driverinit value 202

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Claudiu Manoil <claudiu.manoil@nxp.com>
(forward ported from commit d51c8d3fccb681b878f5b0d731d279cc4fb75e62 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>