LF-12697: support shutdown for for fxls8962 sensor
When system do warm reboot, chip should enter the standby state. It
will ensure in every boot the chip is enabled in same chip internal
logic state.

Add shutdown support for the sensor to disable pm and set the sensor
in standby state in system reboot.

Now this sensor relies on rpm resume suspend to make it enter active
mode and standby mode. When sensor in active mode, system start to
shutdown, because autosuspend_delay value is set 2 seconds, sometimes
this sensor doesn't enter rpm suspend in time when system finished
shutdown. So in next boot the sensor will keep in active mode. In this
case, more time need to be spent to finish the chip reset. According
to datasheet, this sensor need enough time to complete the reset,
otherwise any I2C operation will cause the sensor hang the bus.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Tested-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>