IMXBSP-123-1: arm64: dts: i.mx93: Fix mt9m114 probe random failed
When imx93 use the IO expander for the camera regulators, there is an
issue that is 10% probability reproduced, the mt9m114 camera probed fail.
from the waveform can be found the sequence of the power and clock for
mt9m114 is not correct.

From the mt9m114 reference manual can get the correct order should be:
DOVDD->AVDD->DVDD->EXTCLK

The P3 and P5 are pins of the IO expander pca9538, they are for the EXTCLK
control, so these can't be set regulator always on. That may cause the
EXTCLK signal generated before the power signals of mt9m114. It's not
correct. The EXTCLK enable control should remove from the DTS and should
be controlled in the mt9m114 driver. And also need to enlarge the
"startup-delay-us" and "off-on-delay-us" of EXTCLK to avoid the error
sequence happen.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
(cherry picked from commit de93b2a8c0f7eb2bfc8322600beaa7da1ea13652 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>