LF-12900: vpu: iwave6: Separated encoder parameter for each codec
if using hevc format and setting 5_0 level,
the actural level of encoded stream is 3_0.

the root cause is when open device,the default level is set.
set hevc format first and then set avc format.

the default level of avc format is 50 which equal to 3_0 level for hevc.
when userspace set level of hevc again, the value is not changed.
so driver will use the error value

Separated encoder parameters by codec to prevent
default values of some parameters from being overwritten.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>
Signed-off-by: TaoJiang <tao.jiang_2@nxp.com>
Reviewed-by: MingQian <ming.qian@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>