ILIE-114-1: media: nxp: neo: Perform standard check for VIDIOC_CREATE_BUFS
The check on size has been modified to follow the standard.
The variable size is the minimum size of the plane and
was allocated to sizes[0].
Thus, it was not possible to check sizes[0] < size,
leading to preventing userspace to specify a larger buffer.
This change provides the standard check to perform on size.

Also, the limit of the config node buffer count has been
removed as vb2 core already limits num_buffers to
VB2_MAX_FRAME in vb2_core_queue_init function.

Signed-off-by: Alexi Birlinger <alexi.birlinger@nxp.com>
Reviewed-by: Aymen Sghaier <aymen.sghaier@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>