LF-12915 net: enetc: fix the issue when the mac address changed after suspend/resume
When the MAC address is not provided in U-Boot and cannot be successfully
read from the fuse, if a suspend/resume operation is performed, the MAC
address will be set to a random MAC address. Even though the new MAC
address has been updated in the registers at this point, the MAC
address in the network device (ndev) has not been updated, which may
lead to network communication issues.

Therefore, after setting the new MAC address, update the new MAC address
in the ndev.

Fixes: 61fc886ef26a ("LF-12371 net: enetc: fix suspend/resume issue when eth port down")

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>