LF-12807-6 net: enetc: don't reconfigure if XDP is already enabled
In the previous enetc driver, the Tx and Rx BD rings will be reconfigured
when a bpf program is installed. But it is not necessary if the XDP is
already enabled. So add an optimization for enetc_setup_xdp_prog(), the
BD rings won't be reconfigured and only update bpf program if XDP is
enabled before.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Claudiu Manoil <claudiu.manoil@nxp.com>
(cherry picked from commit f91eb3684ccb5ff30d5cd9098d677e8dfb146c28 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>