phy: introduce C72_LT_INIT API for backplane link training
Currently it is possible that the state of a failed link training
sequence lingers (is preserved) onto the next link training attempt.
That is not okay and we should reset the state so that we always start
from a consistent position.

This is not possible with the current API, because the backplane AN/LT
module actually never explicitly tells the SerDes PHY that a new
training sequence has begun, and we cannot infer that from C72_LT_DONE
means "done successfully". We need to introduce a command specific to
that.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
(cherry picked from commit 82e6150c972313d4336efe212861ebdb9a0ca00a nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>