net: sdk_dpaa: disable multiple technology abilities for backplane
We are not prepared to handle this in sdk_fman's adjust_link() methods,
and this is now the only driver incapable of doing so.

Let the mtip_backplane continue to probe for all backplane modes, but
limit the supported mask at the MAC side.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
(cherry picked from commit 629fed1bb5712bd7e7d2f69d5e37015972e6f765 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>