ANDROID: vendor_hooks: add hook for boost free pages mapped by the exiting process
When the system is under high memory pressure, we observed by trace
that the execution of the exiting process will have a large number of
runnables, which will result in the inability to release the pages
mapped by it's VMA to other apps for use in the system timely.

Therefore, we hope to use this hook interface to release the
physical pages mapped by the exiting process's vma in advance, and
alleviate the tight system memory resources issues.

Bug: 364225494
Bug: 511947074
Bug: 513428566
Change-Id: Id4a7db386d74e9fb8d104e5311a585a4a48811ef
Signed-off-by: Justin Jiang <justinjiang@vivo.corp-partner.google.com>
(cherry picked from commit f58bdcdd1b4b35f71007954bf8708607b8a126f7)