LF-13648-1 ptp: ptp_netc: dynamic adjust oclk_prsc
Before this, the default value of oclk_prsc is 2, or set by dts.
It cannot dynamically meet the provided cycle and pulse width
that are too large or too small by the user.

The current logic will try to set the duty cycle of the output pulse
to 50%.
However, due to IP restrictions, when priv->clk_freq is 333MHz, the
maximum high level time of the pulse is ~6ms. When the user sets a
period longer than 12ms, it cannot guarantee a 50% duty cycle.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>