net: qrtr: bpf_service: Fix rcu logic
RCU locks still require the updating of the object to be protected by
locks or semaphores. Add a new lock to ensure writes to the bpf_filter
pointer cannot race against each other.

Expand the rcu_read_locks to ensure they are held during all places
where the bpf_filter pointer is being used. This will ensure that
references are properly accounted for and can be synchronized on before
attempting to free the filter in the detach ioctl.

Change-Id: I73614522b1eb10652df8688920ac7358e7220d87
Signed-off-by: Chris Lew <quic_clew@quicinc.com>
Signed-off-by: Meenu Raja Sundaram <quic_mrajasun@quicinc.com>