ARM: dts: msm: set mhi buffer length to 8k
When allocate memory for mhi buffer, it needs order-4 memory.
There won't be so many large memory chunk and page allocation
failure will happen on LE projects. Set the mhi buffer length
to 8k so that only order-1 memory is required. Reducing this
size will not impact the message communication with FW.

Change-Id: Ib7e07feae730d20a25ccbe67fb5f7ba9918cbd61
Signed-off-by: Guisen Yang <quic_guiseny@quicinc.com>
CRs-Fixed: 3983268