thermal: imx91: bug fix: Temperature read returns "Resource temporarily unavailable" for values below 0 degree
This issue occurs because the register DATA0[DATA_VAL] is a 16-bit signed value,
but the temperature was not being processed as such before being converted to a 32-bit value.

The fix ensures that the temperature is correctly calculated as a 16-bit signed value
before the conversion to a 32-bit format. This prevents misinterpretation of negative
temperatures and resolves the "Resource temporarily unavailable" error.

Signed-off-by: Yanan Yang <yanan.yang@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>