drivers:iio:stm:accel:st_lis2dw12: rework enable function to remove odr overwriting issue
When events are enabled and at the same time the accelerometer is turned
on and then off, this leads to the cancellation of the ODR needed for
event detection.
This patch adds to the hw structure the variable where the odr requested
by the event management block is saved, this variable is used by the
function that takes care of setting the sensor's odr.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ib27bae4e926f694ee2c9d02bc7ef6ca6b4e73b29