drivers:iio:stm:imu:st_asm330lhhx: fix missing data samples during flush command
The read_fifo function is also used by the flush command
to empty the FIFO. This means that the watermark interrupt
available bit is not necessarily set.
This patch removes the check for the WTM_IA bit to allow
the HW FIFO to be emptied even with the flush command.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Id04c0ae59a94d6b54f6846134e513c7f8023d399