UBUNTU: SAUCE: arm64/dts: rigado-c5k: modify spi irq config
BugLink: https://bugs.launchpad.net/bandon/+bug/2119985 (incorrect spi irq config in device tree (LP: #2119985))

1. spi-max-frequency = <30000000>; -> This frequency is too high and is
not supported by the device, this should be reduced to 20000000

2. interrupts = <8 IRQ_TYPE_LEVEL_LOW>; -> The interrupt type is
incorrect and causes the kernel to reject the irq mapping when the nrc
driver is reloaded. This is seen as the following error: "irq: type
mismatch, failed to map hwirq-8 for gpio@30210000!". This irq definition
in the device tree should be changed to the following:
interrupts = <8 IRQ_TYPE_LEVEL_HIGH>;

Signed-off-by: Bethany Jamison <bethany.jamison@canonical.com>
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>