ANDROID: export symbols for zram module
The zram module is being updated to allow initiating writeback on behalf
of a specific userspace process. This requires the zram module to call
several core kernel functions to access the process's memory map and
walk its page tables.

Export the following symbols so they can be used by the zram module:
 - mm_access: To get a reference to the process's mm_struct.
 - pidfd_get_task: To get the task_struct from a pidfd.
 - walk_page_range: To iterate over the process's page table entries.
 - pte_offset_map_lock: To map and lock a PTE.
 - get_swap_device: To prevent the swapoff/swapon race condition

This change is a prerequisite for adding the new per-process writeback
ioctl to zram.

This fixes the conflict and reland the following commits:
- commit d169319e30af4592a412febf9b7593364b5ef88f
- commit e4e7aec478853bd7338a58703ac4d3e34b7a5ad1
- commit a3654ae7c147bcf458d43931a371acce09aad69a.

Bug: 539739262
Bug: 421065577
Change-Id: Ie1a91d34bfe30b485c3324b5506798e3ee35eb4f
Signed-off-by: Richard Chang <richardycc@google.com>