ANDROID: modpost: add symbol import protection flag to kflagstab
When the unused exports whitelist is provided, the symbol protection bit
is set for symbols not present in the unused exports whitelist.

The flag will be used in the following commit to prevent unsigned
modules from the using symbols other than those explicitly declared by
the such modules ahead of time.

Bug: 519119852
Change-Id: Ibddd628a8b1463397cc95d3d8f556b9bc82bf623
Signed-off-by: Siddharth Nayyar <sidnayyar@google.com>