UPSTREAM: thermal: intel: int340x: Add module parameter to change slider offset
SoC slider value is set by the user (or the default when user has not
modified it). To enhance power efficiency dynamically, the firmware can
optionally auto-adjust the slider value based on the current workload.
This adjustment is governed by an additional parameter known as the
"slider offset". This offset permits the firmware to increase the slider
value up to and including "SoC slider + slider offset".

Add a module parameter to specify this "slier offset" value.

By default, the SoC slider offset is set to 0. This means that SoC is not
allowed to switch slider position.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Link: https://patch.msgid.link/20250825132315.75521-5-srinivas.pandruvada@linux.intel.com
[ rjw: Comment and module param description adjustments ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Bug: 456261880

(cherry picked from commit 8306bcaba06dfa658e3dd852e930b0154cb46c9f)
Change-Id: I2eadc09d148b8960f8e9e5bc932879f7179cd080
Signed-off-by: Yu-cheng Yu <yu-cheng.yu@intel.corp-partner.google.com>