ANDROID: Don't use current pointer for per-process dmabuf accounting
If a child task is created with a partial sharing relationship with the
parent (e.g. CLONE_FILES && !CLONE_VM), AND that child task closes a
dmabuf in the partially shared (files/mm) structure, then the
dma_buf_unaccount_task operation will do nothing since tasks created
with partial sharing relationships do not have associated
task_dma_buf_info accounting structures associated with them. The dmabuf
will be closed, but the accounting information in the parent will still
have a record of the dead buffer. If nothing else happens, the result is
just incorrect accounting.

Later, if a fork is performed the accounting information is copied into
the child, and there a UAF occurs when we dereference the dead dmabuf in
the parent's accounting information in copy_dmabuf_info.

Fix this by tracking task_dma_buf_info in mm_struct and files_struct,
which will be used instead of the current task's task_dma_buf_info
pointer, even for unaccount operations via partial sharing.

Fixes: 857c6e0616d4 ("ANDROID: Track per-process dmabuf RSS")
Bug: 457281015
Bug: 461520624
Bug: 467222928
Change-Id: Idaa4c39de396eb8261c018a816fb549356a2a5cc
Signed-off-by: T.J. Mercier <tjmercier@google.com>