BACKPORT: mempool: add mempool_{alloc,free}_bulk
Add a version of the mempool allocator that works for batch allocations
of multiple objects.  Calling mempool_alloc in a loop is not safe because
it could deadlock if multiple threads are performing such an allocation
at the same time.

As an extra benefit the interface is build so that the same array can be
used for alloc_pages_bulk / release_pages so that at least for page
backed mempools the fast path can use a nice batch optimization.

Note that mempool_alloc_bulk does not take a gfp_mask argument as it
must always be able to sleep and doesn't support any non-trivial
modifiers.  NOFO or NOIO constrainst must be set through the scoped API.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Link: https://patch.msgid.link/20251113084022.1255121-8-hch@lst.de
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>

Bug: 542679931
(cherry picked from commit ac529d86ad26d632d3c70b7c5b839282a3294d2f)
(for "ABI Compliance", reverted change in prototype of mempool_free())
Change-Id: I8c5ad68fd4ddd0ba9d4dd6f6de6be3e9f1c33a92
Signed-off-by: Eric Biggers <ebiggers@google.com>