ANDROID: pkvm: x86: Disable nested feature absent warning for pvIOMMU
For shadow iommu, nested support was necessary as it used the second
level pagetable in nested mode to shadow the host iommu pagetable. This
is no longer a requirement for pviommu as host doesn't own the
pagetables and uses hypercalls to manage iommu. So the warning does not
apply for pviommu.

Bug: 391542119
Upstream-Task: 402758258
Test: Build and Boot

Change-Id: I85e1656820a180f5e1e12ce1506d77d006b434d6
Signed-off-by: Vineeth Pillai <vineethrp@google.com>