ANDROID: ABI: refresh, CRC fixups and allowed breaks
We need to add a new pointer to mm_struct, but there is only 1 ABI
reservation. Use the mm_struct ABI reservation to point to a new
mm_struct_abi_extend which contains a task_dma_buf_info pointer, and can
be extended with other fields in the future.

Since mm_struct_abi_extend is meant to be available for extension in the
future, vendor modules may NOT access this pointer as the size of the
struct cannot be expected to be stable.

Bug: 459757379
Bug: 461520624
Change-Id: I1707f1184c7e97869c9e4eaff0bcdde13b3725df
Signed-off-by: T.J. Mercier <tjmercier@google.com>
(cherry picked from commit 5b6cecf56c7250013e532423509fd9aa77e53456)
Signed-off-by: Lee Jones <joneslee@google.com>