ANDROID: Select rustc-1.91.1 (paired)
This is Rust part of the  compiler pair the toolchain team has matched
LLVM revisions of and produced a release branch for.

For cross-lang helper inlining to work, this needs to be paired with
r584948, but updating clang to that revision tickles several other bugs
at the moment.

This does not include an update to r584948 because that requires further
updates elsewhere.

Bug: 457798339
Change-Id: I629aaea4643720b2327eff7dd7f1807eaebd6fbc
Signed-off-by: Matthew Maurer <mmaurer@google.com>