ANDROID: arch_topology: drop unused vendor declaration
This flag was added in commit 8b49160d2343 ("ANDROID: topology: Add a
flag to indicate topology has been updated"). However, the declaration
in arch_topology.h is not being used in GKI and it causes merge issues.
Drop it from this header file as it can be declared on the vendor side.

Bug: 187234873
Cc: Sai Harshini Nimmala <quic_snimmala@quicinc.com>
Fixes: 8b49160d2343 ("ANDROID: topology: Add a flag to indicate topology has been updated")
Change-Id: I0e7e4a376b47de8ba61827fd46eb71a551a597bc
Signed-off-by: Carlos Llamas <cmllamas@google.com>